dom/media/mediasource/test/test_HaveMetadataUnbufferedSeek_mp4.html

<!DOCTYPE HTML>
<html>
<head>
  <title>MSE: seekable attribute before end of stream</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="mediasource.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

runWithMSE(async (ms, v) => {
  await once(ms, "sourceopen");
  const sb = ms.addSourceBuffer("video/mp4");

  const arrayBuffer = await fetchWithXHR("bipbop/bipbop2s.mp4");
  // 25819 is the offset of the first media segment's end
  sb.appendBuffer(new Uint8Array(arrayBuffer, 0, 25819));

  const target = 1.3;

  await once(v, "loadeddata");
  ok(v.readyState >= v.HAVE_CURRENT_DATA, "readyState is >= CURRENT_DATA");
  v.currentTime = target;

  await once(v, "seeking");
  is(v.readyState, v.HAVE_METADATA);
  // 25819 is the offset of the first media segment's end
  sb.appendBuffer(new Uint8Array(arrayBuffer, 25819));
  await once(sb, "updateend");
  ms.endOfStream();
  await once(v, "seeked");
  SimpleTest.finish();
});

</script>
</pre>
</body>
</html>
